// lib/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck
/*
 * This file exports all enum related types from the schema.
 *
 * 🟢 You can import this file directly.
 */

export const Gender = {
	MALE: "MALE",
	FEMALE: "FEMALE",
} as const;

export type Gender = (typeof Gender)[keyof typeof Gender];

export const OrderStatus = {
	PENDING: "PENDING",
	PROCESSING: "PROCESSING",
	SHIPPED: "SHIPPED",
	DELIVERED: "DELIVERED",
	CANCELLED: "CANCELLED",
} as const;

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const CategoryName = {
	T_SHIRTS: "T_SHIRTS",
	JEANS: "JEANS",
	HOODIES: "HOODIES",
	DRESSES: "DRESSES",
	JACKETS: "JACKETS",
} as const;

export type CategoryName = (typeof CategoryName)[keyof typeof CategoryName];